pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "containerd-py"
version = "0.1.0"
description = "Python reimplementation of containerd control plane with gRPC compatibility"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
  { name = "py-containerd maintainers" }
]
keywords = ["containerd", "oci", "grpc", "containers", "runc"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: POSIX :: Linux",
  "Framework :: AsyncIO",
]
dependencies = [
  "grpcio>=1.63.0",
  "grpcio-tools>=1.63.0",
  "protobuf>=5.27.0",
  "click>=8.1.7",
  "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.8",
  "mypy>=1.10.0",
  "ruff>=0.5.0",
  "types-PyYAML",
]

[project.scripts]
ctr-py = "containerd_py.cli.main:cli"
containerd-py = "containerd_py.daemon:main"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["containerd_py", "containerd_py.cli", "containerd_py.server", "containerd_py.api", "containerd_py.content", "containerd_py.images", "containerd_py.tasks", "containerd_py.state"]

